Refactoring external to leverage naming
